Move single ballot URL and views, redirect
There are some small changes to BallotPaperView in the move from
ConstituencyDetailView, mainly due to taking the Ballot from the URL
directly and not needing to look up the Post.

Otherwise the view is the same: in need of some changes
