+++ ynr/apps/elections/templates/elections/ballot_view.html
@@ -31,17 +31,92 @@
 
 {% block body_class %}constituency{% endblock %}
 
-{% block title %}{% blocktrans with election_name=election_data.name %}Candidates for {{ post_label_shorter }} in the {{ election_name }}{% endblocktrans %}{% endblock %}
+{% block title %}{% blocktrans with election_name=election.name %}Candidates for {{ post_label_shorter }} in the {{ election_name }}{% endblocktrans %}{% endblock %}
 
 {% block hero %}
   <h1>{% blocktrans %}Candidates for <span id="constituency-name">{{ post_label_shorter }}</span>{% endblocktrans %}</h1>
 {% endblock %}
 
+
 {% block content %}
+<div class="row">
+    <div class="columns large-9">
 
-{% include 'candidates/_candidates_for_post.html' %}
+        {% if post_election.cancelled %}
+        <div class="panel">
+          ❌ The poll for this election was cancelled
+        </div>
+        {% endif %}
 
-{% block other_resources %}
-{% endblock %}
+        {% if not candidates_locked %}
+        <div class="panel">
+            <p>These candidates haven't been confirmed by the official "nomination papers"
+                from the council yet. This means they might not all end up on the ballot paper.</p>
+            <p>We will manually verify each candidate when the nomination papers are published.</p>
+        </div>
+        {% endif %}
+
+        {% include 'candidates/_candidates_for_post.html' %}
+
+
+        {# TODO Remove elections that aren't FPTP #}
+        {% if post_election.election.in_past and not post_election.cancelled %}
+        <section>
+        <h3>Results</h3>
+        {% include "uk_results/includes/vote_result_item.html" with post_election=post_election %}
+        </section>
+        {% endif %}
+
+
+
+        {% if not candidates_locked %}
+            {% if some_official_documents %}
+                <a href="{% url "bulk_add_from_sopn" election post_data.id %}" class="button">Add candidates from nomination paper in {{post_data.label}}</a>
+            {% endif %}
+        {% endif %}
+
+        {% if request.user.is_authenticated %}
+          {% if not user_has_suggested_lock %}
+            {% if not candidates_locked %}
+                {% if election.show_official_documents and some_official_documents %}
+                <form method=post id="suggest_lock_form" action="{% url 'constituency-suggest-lock' election_id=election %}">
+                    {% csrf_token %}
+                    <h3>Suggest locking</h3>
+
+
+                    <p>If you think this list is complete, please suggest that we lock it.
+                    Please only do this with good reason, for example if you have reviewed
+                    one of the official lists of candidates ("Nomination papers" or
+                    "Statement of Persons Nominated") linked to above.</p>
+
+                    <p>Locking a list means that candidates can't be added or removed, but
+                    their details can still be edited. This is useful when we know that
+                    the list is final.</p>
+
+
+                    {{ suggest_lock_form.postextraelection }}
+
+                    {{ suggest_lock_form.justification.label_tag }}
+                    <label for="{{ suggest_lock_form.justification.id_for_label }}">{{ suggest_lock_form.justification.help_text }}</label>
+                    {{ suggest_lock_form.justification }}
+                    <button type="submit" class="button small secondary">Suggest locking</button>
+                </form>
+                {% endif %}
+            {% endif %}
+          {% else %}
+            <h3>Suggest locking</h3>
+            {% if candidates_locked %}
+            <p>Thanks! You suggested locking this list, and we did!</p>
+            {% else %}
+            <p>Thanks, you've suggested we lock this list – we'll review it soon.</p>
+            {% endif %}
+          {% endif %}
+        {% endif %}
+
+    </div>
+    <div class="columns large-3">
+        {% include "uk/data_timeline.html" %}
+    </div>
+</div>
 
 {% endblock %}
